+++ packages/map-template/src/components/Directions/Directions.jsx
@@ -19,6 +19,8 @@ import substepsToggledState from "../../atoms/substepsToggledState";
 import getDesktopPaddingLeft from "../../helpers/GetDesktopPaddingLeft";
 import getMobilePaddingBottom from "../../helpers/GetMobilePaddingBottom";
 import distanceUnitSystemSelector from '../../selectors/distanceUnitSystemSelector';
+import getDesktopPaddingBottom from "../../helpers/GetDesktopPaddingBottom";
+import kioskLocationState from "../../atoms/kioskLocationState";
 
 let directionsRenderer;
 
@@ -57,6 +59,8 @@ function Directions({ isOpen, onBack, onSetSize, snapPointSwiped }) {
 
     const [substepsOpen, setSubstepsOpen] = useRecoilState(substepsToggledState);
 
+    const kioskLocation = useRecoilValue(kioskLocationState)
+
     const isDesktop = useMediaQuery('(min-width: 992px)');
 
     useEffect(() => {
@@ -77,8 +81,8 @@ function Directions({ isOpen, onBack, onSetSize, snapPointSwiped }) {
                 mapsIndoors: mapsIndoorsInstance,
                 fitBoundsPadding: {
                     top: padding,
-                    bottom: isDesktop ? padding : getMobilePaddingBottom(),
-                    left: isDesktop ? getDesktopPaddingLeft() : padding,
+                    bottom: getBottomPadding(padding),
+                    left: getLeftPadding(padding),
                     right: padding
                 }
             });
@@ -99,6 +103,39 @@ function Directions({ isOpen, onBack, onSetSize, snapPointSwiped }) {
             }
         }
     }, [isOpen, directions, mapsIndoorsInstance, travelMode]);
+
+
+    /**
+     * Get bottom padding when getting directions.
+     * Calculate all cases depending on the kioskLocation id prop as well.
+     */
+    function getBottomPadding(padding) {
+        if (isDesktop) {
+            if (kioskLocation) {
+                return getDesktopPaddingBottom();
+            } else {
+                return padding;
+            }
+        } else {
+            return getMobilePaddingBottom();
+        }
+    }
+
+    /**
+     * Get left padding when getting directions.
+     * Calculate all cases depending on the kioskLocation id prop as well.
+     */
+    function getLeftPadding(padding) {
+        if (isDesktop) {
+            if (kioskLocation) {
+                return padding;
+            } else {
+                return getDesktopPaddingLeft();
+            }
+        } else {
+            return padding;
+        }
+    }
 
     /*
      * Make sure directions stop rendering on the map when the Directions view is not active anymore.