+++ packages/map-template/src/components/Search/Search.jsx
@@ -17,6 +17,9 @@ import fitBoundsLocation from '../../helpers/fitBoundsLocation';
 import getDesktopPaddingLeft from '../../helpers/GetDesktopPaddingLeft';
 import languageState from '../../atoms/languageState';
 import { useTranslation } from 'react-i18next';
+import kioskLocationState from '../../atoms/kioskLocationState';
+import getDesktopPaddingBottom from '../../helpers/GetDesktopPaddingBottom';
+import getMobilePaddingBottom from '../../helpers/GetMobilePaddingBottom';
 
 /**
  * Show the search results.
@@ -68,6 +71,8 @@ function Search({ onSetSize }) {
     const currentLanguage = useRecoilValue(languageState);
 
     const isDesktop = useMediaQuery('(min-width: 992px)');
+
+    const kioskLocation = useRecoilValue(kioskLocationState);
 
     /**
      * Get the locations and filter through them based on categories selected.
@@ -194,7 +199,39 @@ function Search({ onSetSize }) {
             mapsIndoorsInstance.setFloor(locationFloor);
         }
 
-        fitBoundsLocation(location, mapsIndoorsInstance, isDesktop ? 0 : 200, isDesktop ? getDesktopPaddingLeft() : 0)
+        fitBoundsLocation(location, mapsIndoorsInstance, getBottomPadding(), getLeftPadding());
+    }
+
+    /**
+     * Get bottom padding when selecting a location.
+     * Calculate all cases depending on the kioskLocation id prop as well.
+     */
+    function getBottomPadding() {
+        if (isDesktop) {
+            if (kioskLocation) {
+                return getDesktopPaddingBottom();
+            } else {
+                return 0;
+            }
+        } else {
+            return 200;
+        }
+    }
+
+    /**
+     * Get left padding when selecting a location.
+     * Calculate all cases depending on the kioskLocation id prop as well.
+     */
+    function getLeftPadding() {
+        if (isDesktop) {
+            if (kioskLocation) {
+                return 0;
+            } else {
+                return getDesktopPaddingLeft();
+            }
+        } else {
+            return 0;
+        }
     }
 
     /*