+++ packages/map-template/src/components/Sidebar/Sidebar.jsx
@@ -1,5 +1,5 @@
 import { useEffect } from 'react';
-import { useRecoilState } from 'recoil';
+import { useRecoilState, useRecoilValue } from 'recoil';
 import currentLocationState from '../../atoms/currentLocationState';
 import filteredLocationsByExternalIDState from '../../atoms/filteredLocationsByExternalIDState';
 import Modal from './Modal/Modal';
@@ -8,6 +8,8 @@ import Wayfinding from '../Wayfinding/Wayfinding';
 import Directions from '../Directions/Directions';
 import Search from '../Search/Search';
 import LocationsList from '../LocationsList/LocationsList';
+import locationIdState from '../../atoms/locationIdState';
+import kioskLocationState from '../../atoms/kioskLocationState';
 
 /**
  * @param {Object} props
@@ -22,6 +24,8 @@ import LocationsList from '../LocationsList/LocationsList';
 function Sidebar({ directionsFromLocation, directionsToLocation, pushAppView, currentAppView, appViews }) {
     const [currentLocation, setCurrentLocation] = useRecoilState(currentLocationState);
     const [filteredLocationsByExternalIDs, setFilteredLocationsByExternalID] = useRecoilState(filteredLocationsByExternalIDState);
+    const [, setLocationId] = useRecoilState(locationIdState);
+    const kioskLocation = useRecoilValue(kioskLocationState)
 
     /*
      * React on changes on the current location and directions locations and set relevant bottom sheet.
@@ -33,8 +37,12 @@ function Sidebar({ directionsFromLocation, directionsToLocation, pushAppView, cu
             pushAppView(appViews.WAYFINDING);
         } else if (currentLocation) {
             pushAppView(appViews.LOCATION_DETAILS, currentLocation);
-        } else if (filteredLocationsByExternalIDs?.length > 0) {
+        } else if (filteredLocationsByExternalIDs?.length > 1) {
             pushAppView(appViews.EXTERNALIDS);
+            // If there is only one external ID, behave the same as having the location ID prop. 
+        } else if (filteredLocationsByExternalIDs?.length === 1) {
+            setCurrentLocation(filteredLocationsByExternalIDs[0])
+            setLocationId(filteredLocationsByExternalIDs[0].id)
         } else {
             pushAppView(appViews.SEARCH);
         }
@@ -44,9 +52,13 @@ function Sidebar({ directionsFromLocation, directionsToLocation, pushAppView, cu
      * Close the location details page and navigate to either the Locations list page or the Search page.
      */
     function closeLocationDetails() {
-        if (filteredLocationsByExternalIDs?.length > 0) {
+        if (filteredLocationsByExternalIDs?.length > 1) {
             pushAppView(appViews.EXTERNALIDS);
             setCurrentLocation();
+        } else if (filteredLocationsByExternalIDs?.length === 1) {
+            pushAppView(appViews.SEARCH);
+            setCurrentLocation();
+            setFilteredLocationsByExternalID([]);
         } else {
             pushAppView(appViews.SEARCH);
             setCurrentLocation();
@@ -62,9 +74,20 @@ function Sidebar({ directionsFromLocation, directionsToLocation, pushAppView, cu
         setFilteredLocationsByExternalID([]);
     }
 
+    /**
+     * Close the Directions page and navigate to the different pages based on the kioskLocation.
+     */
+    function closeDirections() {
+        if (kioskLocation) {
+            pushAppView(appViews.LOCATION_DETAILS)
+        } else {
+            pushAppView(appViews.WAYFINDING)
+        }
+    }
+
     const pages = [
         <Modal isOpen={currentAppView === appViews.SEARCH} key="A">
-            <Search/>
+            <Search />
         </Modal>,
         <Modal isOpen={currentAppView === appViews.EXTERNALIDS} key="B">
             <LocationsList
@@ -78,6 +101,7 @@ function Sidebar({ directionsFromLocation, directionsToLocation, pushAppView, cu
             <LocationDetails
                 onStartWayfinding={() => pushAppView(appViews.WAYFINDING)}
                 onBack={() => closeLocationDetails()}
+                onStartDirections={() => pushAppView(appViews.DIRECTIONS)}
             />
         </Modal>,
         <Modal isOpen={currentAppView === appViews.WAYFINDING} key="D">
@@ -92,7 +116,7 @@ function Sidebar({ directionsFromLocation, directionsToLocation, pushAppView, cu
         <Modal isOpen={currentAppView === appViews.DIRECTIONS} key="E">
             <Directions
                 isOpen={currentAppView === appViews.DIRECTIONS}
-                onBack={() => pushAppView(appViews.WAYFINDING)}
+                onBack={() => closeDirections()}
             />
         </Modal>
     ]