+++ packages/map-template/CHANGELOG.md
@@ -5,6 +5,24 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.24.0] - 2023-12-11
+
+### Added
+
+- New prop, `useMapProviderModule`: When set to true, the modules in the MapsIndoors Solution will be used to determine which map type to use.
+
+## [1.23.0] - 2023-12-06
+
+### Added
+
+- Add support showing directions immediately after selecting a location in a Kiosk mode.
+
+## [1.22.0] - 2023-12-06
+
+### Added
+
+- Add support for having one external ID.
+
 ## [1.21.0] - 2023-11-30
 
 ### Added